Split CreateSessionTests.cs into focused test files per class
Agent-Logs-Url: https://github.com/ECU-Pirate-Forge/pulse-2.0/sessions/841c65a0-f175-4c58-af77-0eb8bdc18e19
